Add count-inversions implementation skeleton
Non-functioning code skeleton for the count-inversions algorithm.
Count-inversions is an algorithm for counting the number of inversions
(instances where the value at an index n is greater than the value at
a higher index) in an array. It uses the divide-and-conquer paradigm
to first recursively count the number of inversions in each half of
the array, then count the inversions between halves.