Make toInt operate on entire number
In order to carry digits in addition, the numbers involved had to
be stored as integer arrays from the start, rather than converting
individual characters to ints as the algorithm went along. Now
toInt takes in the numbers to be mulitplied as strings and returns
a pointer to an int array.