Implement recursive multiplication for some inputs
It works... but only on inputs with an equal number of digits,
and the number of digits must be a power of 2. Basically it
can multiply two 2-digit or two 4-digit numbers.

Removed the ill-conceived getabcd function and replaced with
several new functions. multiply() is the recursive function which
obtains a, b, c, and d from a helper function splitIntArray,
generates the 3 terms of the multiplication equation, and adds them
together to produce the final product. ArrToInt() was added as a
helper function for the base case of multiply, when one of the
terms reaches one digits and the two terms are covnerted to ints
and multiplied to end the recursion.