Declare function and test array for quicksort
Quicksort is a O(nlogn), divide-and-conquer sorting algorithm. I
operates by selecting a 'partition' element (ideally at random)
and swapping elements in the array so that all elements less than
the partition are moved to the left of it, and all elements greater
than the partition are moved to the right of it. Then the left and
right parts are themselves 'sorted' recursively by the same method.

This commit does none of that yet. It just declares the sort
function and sets up a test array.