Add length parameter to sort and remove semicolons
The sort function requires a length parameter, so the function
declaration and the initial call were modified to reflect that.
A length of just 1 element represents the base case of the
recursion, so the function simply returns in this case.

Also I forgot how python works for a minute there and included a
bunch of unecessary semicolons. Those are gone now