Add partition step and swap helper function
The data is partitioned iterating over the array and moving any
element less than the pivot value to the left part of the array.
This is done using an additional variable i that represents the
index of the smallest value greater than the pivot - any value
less than the pivot is swapped at the value at this index, which
is then incremented.

This version still uses a naive choice of pivot.