Test quicksort with sorted and unsorted inputs
Using the first element as the pivot works fine when the array is
random, but is bad when the array is already sorted. To see this
for myself I changed quicksort to run sequentially on first a
random, then a sorted array, and adjusted quicksort_test to allow
for greater recursion. As expected, sorted arrays run in O(n^2)
time - not good.