Removed redundant sorting of the pivot
After partitioning in quicksort, the pivot value ends up in exactly
the 'right' spot (the spot in the array that it will have once the
array is sorted). There is no need to include it in either part
when recursing, then, so now it's just concatenated in at the end.