Refactor recursion step and strengthen checking
Added conditionals to the recursion step to check for empty or one-
element arrays; in those cases the array is not recursed, just
added on at the end.

In doing this I found that I accidentally removed elements from
the array and this was not picked up by the correctness checking
function; it now makes sure that the final array has the same
length as the starting array.