Remove unnecessary parameters from sort()
The start parameter in sort() turned out to always be 0, so there
is no point in passing it to every iteration of the function.
Similarly, length can be simply retrieved by len(arr) rather than
calculated and passed in to every recursion. Removing these
parameters also gives the sort() function the same parameters as
mergesort, which should allow for a streamlined test script.