Only check correctness if instructed
Added a 'check' parameter which is given the value True if any
command line argument is given other than the array length.
The main function now only checks for correctness with quicksort if
this check parameter is ture; otherwise it simply runs random_select
and reports the time it took, allowing the testing of much larger
arrays (up to ~64000000 elements in under a minute).